requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-anthropic>=0.2.0
anthropic>=0.40.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
newsapi-python>=0.2.7
feedparser>=6.0.10
apscheduler>=3.10.0
twilio>=9.0.0
flask>=3.0.0
python-dotenv>=1.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
ragas>=0.2.0
datasets>=2.18.0
